+++ lib/linzer/message/adapter/http_gem/common.rb
@@ -20,14 +20,6 @@ module Linzer
             @operation.headers[name]
           end
 
-          # Attaches a signature to the response.
-          # @param signature [Signature] The signature to attach
-          # @return [Object] The underlying response object
-          def attach!(signature)
-            signature.to_h.each { |h, v| @operation.headers[h] = v }
-            @operation
-          end
-
           private
 
           # Retrieves an HTTP field value from the request or response headers.
@@ -40,6 +32,17 @@ module Linzer
             return nil if has_tr # XXX: is there a library actually supporting trailers?
             value = @operation.headers[name.value.to_s]
             value.dup&.strip
+          end
+
+          # Sets a header on the underlying HTTP message.
+          #
+          # If a header with the given name already exists, its value is overwritten.
+          #
+          # @param header [String] the header name
+          # @param value [String] the header value
+          # @return [String] the value assigned to the header
+          def set_header!(header, value)
+            @operation.headers[header] = value
           end
         end
       end